refactor(function-tree): refactor based on comments
