+++ packages/function-tree/providers.js
@@ -2,4 +2,4 @@ exports.ContextProvider = require('./lib/providers/Context').default
 exports.ExecutionProvider = require('./lib/providers/Execution').default
 exports.PropsProvider = require('./lib/providers/Props').default
 exports.PathProvider = require('./lib/providers/Path').default
-exports.PathProvider = require('./lib/providers/StopExecution').default
+exports.StopExecutionProvider = require('./lib/providers/StopExecution').default
